addressed change requests
